% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_data.R
\name{aggregate_path_metrics}
\alias{aggregate_path_metrics}
\title{Construct a data.frame with aggregate study data}
\usage{
aggregate_path_metrics(
  graph_obj,
  analysis_obj,
  parsed_responses,
  instruments_df
)
}
\arguments{
\item{graph_obj}{iGraph The object returned in \code{paths_to_graph(...)@graph@g})}

\item{analysis_obj}{The object returned by \code{wikidata_analysis()}}

\item{parsed_responses}{The result of \code{parse_json()} list-applied to all responses}

\item{instruments_df}{data.frame of seed terms and Wikidata codes, like RAWINSTRUMENTS}
}
\description{
Construct a data.frame with aggregate study data
}
