% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondence_gplot.R
\name{ca_gplot}
\alias{ca_gplot}
\title{Plot Correspondence Analysis}
\usage{
ca_gplot(unified_data_obj)
}
\arguments{
\item{unified_data_object}{The object created by \code{aggregate_path_metrics()}}
}
\value{
data data.frame linking node names with wikidata codes and CA row coordinates

ca_obj The result of ca::ca()

ca_plot_obj The object created by ca::plot.ca()

index An index linking node names to wikidata codes, without the whole URI

ca_row_coord CA row coordinates

ca_col_coord CA col coordinates

plotly_rows Row coordinates used in plotly (labels, jittered)

plotly_cols Column coordinates used in plotly (labels, jittered)
}
\description{
Plot Correspondence Analysis
}
